improve quality

+++ ReadingIsGood.Application/Commands/Book/BookCreateCommand.cs
@@ -7,5 +7,5 @@ public class BookCreateCommand : IRequest<BookResponse>
 {
     public string Title { get; set; } = "";
     public int StockQuantity { get; set; }
-    public double BookPrice { get; set; }
+    public decimal BookPrice { get; set; }
 }
