+++ ReadingIsGood.Application/Responses/BookResponse.cs
@@ -5,5 +5,5 @@ public class BookResponse : BaseResponse
     public int Id { get; set; }
     public string Title { get; set; } = "Unknown Title";
     public int StockQuantity { get; set; }
-    public double BookPrice { get; set; }
+    public decimal BookPrice { get; set; }
 }
