+++ ReadingIsGood.Application/Responses/CustomerStatisticsResponse.cs
@@ -8,5 +8,5 @@ public class CustomerStatisticsResponse : BaseResponse
     public string Month { get; set; } = "";
     public int TotalOrderCount { get; set; }
     public int TotalBookCount { get; set; }
-    public double TotalPurchasedAmount { get; set; }
+    public decimal TotalPurchasedAmount { get; set; }
 }
