+++ ReadingIsGood.Domain/Entities/Book.cs
@@ -6,5 +6,5 @@ public class Book : Entity
 {
     public string Title { get; set; } = "";
     public int StockQuantity { get; set; }
-    public double BookPrice { get; set; }
+    public decimal BookPrice { get; set; }
 }
